Removed warnings

+++ DemoApplication/iPhone/TableViewCells/CategoryPTableViewCell_iPhone.m
@@ -36,7 +36,7 @@
 
 - (IBAction)selectCellView:(UIButton *)button {
     
-    [parentController openMediaInfoByIndex:index + button.tag];
+    [parentController openMediaInfoByIndex:index + (int)button.tag];
     
 }
